Setup improved DynamicDisplayList
* Add BaseDisplayList as the new untyped display list class.
* Add a type parameter to DynamicDisplayList.
* Add a Populate() method to DynamicDisplayList.
* Add abstract InstantiateElement() method to DynamicDisplayList.
