+++ Runtime/DynamicDisplayList.cs
@@ -1,155 +1,28 @@
 using System;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using UnityEngine;
 
-public class DynamicDisplayList : MonoBehaviour
+public abstract class DynamicDisplayList<D> : BaseDisplayList
 {
-    [SerializeField]
-    private Transform _root;
+    private List<D> _data = null;
 
-    [SerializeField]
-    private bool _forwardDragElements;
-
-    [SerializeField]
-    private bool _invertedDirection;
-
-    private List<Transform> _elements = new List<Transform>();
-
-    public Transform Root
+    public ReadOnlyCollection<D> Data
     {
-        get { return _root; }
+        get { return _data.AsReadOnly(); }
     }
 
-    public bool IsInverted
+    public virtual void Populate(List<D> data)
     {
-        get { return _invertedDirection; }
-    }
+        _data = data ?? throw new ArgumentNullException("data parameter must not be null");
 
-    public IEnumerable<Transform> Elements
-    {
-        get { return _elements; }
-    }
+        Clear();
 
-    public Transform this[int key]
-    {
-        get { return _elements[key]; }
-    }
-
-    public void Clear()
-    {
-        foreach (var element in _elements)
+        foreach (var dataElement in _data)
         {
-            Destroy(element.gameObject);
-        }
-        _elements.Clear();
-    }
-
-    public T CreateChild<T>(
-        T prefab,
-        bool setToIdentity = true,
-        Vector2? sizeDelta = null)
-    where T : Component
-    {
-        var instance = Instantiate(prefab, _root);
-
-        var t = instance.transform;
-        if (setToIdentity)
-        {
-            t.localPosition = Vector3.zero;
-            t.localRotation = Quaternion.identity;
-            t.localScale = Vector3.one;
-        }
-
-        if (t is RectTransform rect)
-        {
-            if (setToIdentity)
-            {
-                rect.anchoredPosition = Vector2.zero;
-            }
-
-            if (sizeDelta.HasValue)
-            {
-                rect.sizeDelta = sizeDelta.Value;
-            }
-        }
-
-        return instance;
-    }
-
-    public T CreateAndAppendChild<T>(
-        T prefab,
-        bool setToIdentity = true,
-        Vector2? sizeDelta = null)
-    where T : Component
-    {
-        var instance = CreateChild(prefab, setToIdentity, sizeDelta);
-        AppendElement(instance.transform);
-        return instance;
-    }
-
-    public void AppendElement(Transform element)
-    {
-        _elements.Add(element);
-        element.gameObject.SetActive(true);
-    }
-
-    /// <summary>
-    /// Inserts the given element at the specified index.
-    /// </summary>
-    ///
-    /// <param name="index"></param>
-    /// <param name="element"></param>
-    ///
-    /// <exception cref="ArgumentOutOfRangeException">
-    /// <paramref name="index"/> is less than 0, or <paramref name="index"/> is greater than Count.
-    /// </exception>
-    public void Insert(int index, Transform element)
-    {
-        if (index < 0 || index > _elements.Count)
-        {
-            throw new ArgumentOutOfRangeException();
-        }
-
-        if (index == _elements.Count)
-        {
-            AppendElement(element);
-            return;
-        }
-
-        element.SetSiblingIndex(index);
-        _elements.Insert(index, element);
-
-        if (element == null)
-        {
-            Debug.LogError(
-                $"Element could not be inserted into List {_root.name} at {index}, " +
-                $"Data: {element}");
-        }
-        else
-        {
-            element.gameObject.SetActive(true);
+            InstantiateElement(dataElement);
         }
     }
 
-    public bool RemoveAt(int index)
-    {
-        if (index < 0 || index >= _elements.Count)
-        {
-            return false;
-        }
-
-        var element = _elements[index];
-        _elements.RemoveAt(index);
-        element.gameObject.transform.SetSiblingIndex(_elements.Count);
-        element.gameObject.SetActive(false);
-
-        _elements.Add(element);
-
-        return true;
-    }
-
-    private void Reset()
-    {
-        _root = transform;
-    }
+    protected abstract void InstantiateElement(D data);
 }
