+++ Runtime/IDisplayElement.cs
@@ -2,8 +2,8 @@
 /// Common interface for all UI elements that can be displayed in a <see cref="DisplayList{T, U}"/>.
 /// </summary>
 ///
-/// <typeparam name="T">The data type used to populate this display element.</typeparam>
-public interface IDisplayElement<T>
+/// <typeparam name="D">The data type used to populate this display element.</typeparam>
+public interface IDisplayElement<D>
 {
-    void Populate(T data);
+    void Populate(D data);
 }
